@@ -1,6 +1,6 @@
 [package]
 description = "Server for swhkdp to run shell commands at invoking user level."
-edition = "2021"
+edition = "2024"
 license = "BSD-2-Clause"
 name = "swhks"
 version = "1.3.1-git"
