@@ -25,13 +25,13 @@ impl Env {
                     Self::pkexec_err();
                 }
                 EnvError::GenericError(e) => {
-                    log::error!("Error: {}", e);
+                    log::error!("Error: {e}");
                     Self::pkexec_err();
                 }
             },
         };
         let config_folder_location = PathBuf::from("/etc");
-        let runtime_dir = PathBuf::from(format!("/run/user/{}", pkexec_id));
+        let runtime_dir = PathBuf::from(format!("/run/user/{pkexec_id}"));
 
         Self { pkexec_id, config_folder_location, runtime_dir }
     }
